// src/modules/aigw/backend/types/aigw-partner.types.ts
// Auto-generated by Codegen Engine
export interface AigwPartnerDO {
  id: string
  partner_code: string
  name: string
  level: string
  registered_capital: number
  credit_code: string
  contact_name: string
  contact_phone: string
  region?: string
  commission_rate: number
  promo_code: string
  balance: number
  total_commission: number
  status: string
  tenant_id: string
  create_time: string
}

export interface AigwPartnerVO {
  id: string
  partner_code: string
  name: string
  level: string
  registered_capital: number
  credit_code: string
  contact_name: string
  contact_phone: string
  region?: string
  commission_rate: number
  promo_code: string
  balance: number
  total_commission: number
  status: string
  tenant_id: string
  create_time: string
}

export interface AigwPartnerCreateDTO {
  partner_code: string
  name: string
  level: string
  registered_capital: number
  credit_code: string
  contact_name: string
  contact_phone: string
  region?: string
  commission_rate: number
  promo_code: string
  balance: number
  total_commission: number
  status: string
}

export interface AigwPartnerUpdateDTO extends Partial<AigwPartnerCreateDTO> {
  id: string
}

export interface AigwPartnerPageQuery {
  page?: number
  pageSize?: number
  partner_code?: string
  name?: string
  credit_code?: string
  promo_code?: string
}